postTutorial

+++ src/main/java/com/CrudTutorial/Repositories/ITutorialRepository.java
@@ -3,5 +3,5 @@ package com.CrudTutorial.Repositories;
 import com.CrudTutorial.models.Tutorial;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ITutorialRepository extends CrudRepository <Tutorial, Long> {
+public interface ITutorialRepository extends CrudRepository <Tutorial, Integer> {
 }
